Moved Link stuff from Domain, Data and Services to Links folder
