+++ LivingMessiah.Web/Features/NavigationSearch/Index.razor
@@ -1,7 +1,7 @@
 ﻿@page "/NavigationSearch"
 @using Page = LivingMessiah.Web.Links.NavigationSearch
-@using LivingMessiah.Web.Domain
-@inject Services.ILinkService? LinkService
+@using LivingMessiah.Web.Links
+@inject Links.ILinkService? LinkService
 
 <PageTitle>@Page.Title</PageTitle>
