+++ LivingMessiah.Web/Features/SampleCode/Syncfusion_SfDropDownList/BibleBookChapterDDL.razor.cs
@@ -1,7 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using LivingMessiah.Web.Enums;
-using LivingMessiah.Web.Services;
 using Microsoft.AspNetCore.Components;
 
 namespace LivingMessiah.Web.Features.SampleCode.Syncfusion_SfDropDownList;
