+++ LivingMessiah.Web/ServiceCollectionExtensions.cs
@@ -17,6 +17,8 @@ using FluentValidation;
 using LivingMessiah.Web.Features.Parasha.Services;
 using LivingMessiah.Web.Features.Parasha.Data;
 using LivingMessiah.Web.Features.Admin.Wirecast;
+using LivingMessiah.Web.Links;
+using LivingMessiah.Web.Features.SampleCode.Syncfusion_SfDropDownList;
 
 namespace LivingMessiah.Web;
